+++ Source/DiabloUI/dialogs.cpp
@@ -14,7 +14,9 @@
 #include "engine/load_clx.hpp"
 #include "engine/load_pcx.hpp"
 #include "engine/palette.h"
+#include "headless_mode.hpp"
 #include "hwcursor.hpp"
+#include "init.h"
 #include "utils/display.h"
 #include "utils/is_of.hpp"
 #include "utils/language.h"
