+++ Source/control.cpp
@@ -29,7 +29,6 @@
 #include "engine/trn.hpp"
 #include "gamemenu.h"
 #include "headless_mode.hpp"
-#include "init.h"
 #include "inv.h"
 #include "inv_iterators.hpp"
 #include "levels/setmaps.h"
