+++ Source/engine/dx.cpp
@@ -11,6 +11,7 @@
 #include "controls/plrctrls.h"
 #include "engine/render/primitive_render.hpp"
 #include "headless_mode.hpp"
+#include "init.h"
 #include "options.h"
 #include "utils/display.h"
 #include "utils/log.hpp"
