+++ Source/engine/path.h
@@ -6,14 +6,12 @@
 #pragma once
 
 #include <cstdint>
-#include <limits>
 #include <optional>
 
-#include <SDL.h>
 #include <function_ref.hpp>
 
-#include "engine/direction.hpp"
 #include "engine/point.hpp"
+#include "utils/attributes.h"
 
 namespace devilution {
