+++ Source/engine/render/clx_render.cpp
@@ -6,9 +6,11 @@
 #include "clx_render.hpp"
 
 #include <algorithm>
+#include <cstdint>
 
 #include "engine/point.hpp"
 #include "engine/render/blit_impl.hpp"
+#include "engine/surface.hpp"
 #include "utils/attributes.h"
 #include "utils/clx_decode.hpp"
 #include "utils/static_vector.hpp"
