+++ Source/gamemenu.cpp
@@ -13,8 +13,8 @@
 #include "engine/sound_defs.hpp"
 #include "gmenu.h"
 #include "headless_mode.hpp"
-#include "init.h"
 #include "loadsave.h"
+#include "multi.h"
 #include "options.h"
 #include "pfile.h"
 #include "qol/floatingnumbers.h"
