+++ Source/help.cpp
@@ -10,7 +10,7 @@
 #include "DiabloUI/ui_flags.hpp"
 #include "engine/render/clx_render.hpp"
 #include "engine/render/text_render.hpp"
-#include "init.h"
+#include "game_mode.hpp"
 #include "minitext.h"
 #include "qol/chatlog.h"
 #include "stores.h"
