+++ Source/interfac.cpp
@@ -20,10 +20,11 @@
 #include "engine/palette.h"
 #include "engine/render/clx_render.hpp"
 #include "engine/render/primitive_render.hpp"
+#include "game_mode.hpp"
 #include "headless_mode.hpp"
 #include "hwcursor.hpp"
-#include "init.h"
 #include "loadsave.h"
+#include "multi.h"
 #include "pfile.h"
 #include "plrmsg.h"
 #include "utils/log.hpp"
