+++ Source/items.cpp
@@ -27,8 +27,8 @@
 #include "engine/render/clx_render.hpp"
 #include "engine/render/primitive_render.hpp"
 #include "engine/render/text_render.hpp"
+#include "game_mode.hpp"
 #include "headless_mode.hpp"
-#include "init.h"
 #include "inv_iterators.hpp"
 #include "levels/town.h"
 #include "lighting.h"
