+++ Source/levels/gendung.cpp
@@ -15,7 +15,8 @@
 #include "engine/load_file.hpp"
 #include "engine/random.hpp"
 #include "engine/world_tile.hpp"
-#include "init.h"
+#include "game_mode.hpp"
+#include "items.h"
 #include "levels/drlg_l1.h"
 #include "levels/drlg_l2.h"
 #include "levels/drlg_l3.h"
@@ -23,7 +24,9 @@
 #include "levels/reencode_dun_cels.hpp"
 #include "levels/town.h"
 #include "lighting.h"
-#include "options.h"
+#include "monster.h"
+#include "objects.h"
+#include "utils/algorithm/container.hpp"
 #include "utils/bitset2d.hpp"
 #include "utils/is_of.hpp"
 #include "utils/log.hpp"
