+++ Source/lighting.h
@@ -7,12 +7,8 @@
 
 #include <array>
 #include <cstdint>
-#include <optional>
-#include <type_traits>
-#include <vector>
 
 #include <expected.hpp>
-#include <function_ref.hpp>
 
 #include "automap.h"
 #include "engine/displacement.hpp"
