+++ Source/missiles.cpp
@@ -22,8 +22,8 @@
 #include "engine/points_in_rectangle_range.hpp"
 #include "engine/random.hpp"
 #include "engine/render/primitive_render.hpp"
+#include "game_mode.hpp"
 #include "headless_mode.hpp"
-#include "init.h"
 #include "inv.h"
 #include "levels/dun_tile.hpp"
 #include "levels/trigs.h"
