+++ Source/pack.cpp
@@ -9,7 +9,6 @@
 
 #include "engine/random.hpp"
 #include "game_mode.hpp"
-#include "init.h"
 #include "items/validation.h"
 #include "loadsave.h"
 #include "playerdat.hpp"
