+++ Source/stores.cpp
@@ -14,14 +14,14 @@
 #include "controls/plrctrls.h"
 #include "cursor.h"
 #include "engine/backbuffer_state.hpp"
-#include "engine/load_cel.hpp"
 #include "engine/random.hpp"
 #include "engine/render/clx_render.hpp"
 #include "engine/render/primitive_render.hpp"
 #include "engine/render/text_render.hpp"
 #include "engine/trn.hpp"
-#include "init.h"
+#include "game_mode.hpp"
 #include "minitext.h"
+#include "multi.h"
 #include "options.h"
 #include "panels/info_box.hpp"
 #include "qol/stash.h"
