+++ Source/utils/log.hpp
@@ -4,6 +4,7 @@
 
 #include <SDL.h>
 #include <fmt/core.h>
+#include <fmt/format.h>
 #include <fmt/ranges.h>
 
 #include "utils/str_cat.hpp"
