+++ test/language_for_testing.cpp
@@ -1,5 +1,6 @@
 #include <string>
 #include <string_view>
+#include <vector>
 
 std::string_view GetLanguageCode() { return "en"; }
 bool HasTranslation(const std::string &locale) { return true; }
@@ -7,3 +8,4 @@ void LanguageInitialize() { }
 std::string_view LanguageTranslate(const char *key) { return key; }
 std::string_view LanguagePluralTranslate(const char *singular, std::string_view plural, int count) { return count == 1 ? singular : plural; }
 std::string_view LanguageParticularTranslate(std::string_view context, std::string_view message) { return message; }
+std::vector<std::string> GetLocales() { return {}; }
