+++ test/timedemo_test.cpp
@@ -2,10 +2,11 @@
 #include <gtest/gtest.h>
 #include <iostream>
 
-#include "diablo.h"
+#include "engine/assets.hpp"
 #include "engine/demomode.h"
 #include "game_mode.hpp"
 #include "headless_mode.hpp"
+#include "init.h"
 #include "lua/lua.hpp"
 #include "monstdat.h"
 #include "options.h"
